Move core modules out of scionlab.util, cleanup
Moved modules certificates, openvpn_config, generate and config_tar from
scionlab.util to scionlab.
Renamed scionlab.util.generate to scionlab.scion_config.

Removed some unused files.
